Ensure that exit code is set properly
Currently both tools return 0 if problems occur but they do not
ultimately have an uncaught exception. For example, if metaImport.py
calls main_validate() and gets return code != 0, it will exit with
success despite printing an error message. validateData.py does the same
when run standalone.

This PR calls sys.exit() with the return value of the validation. In the
case that the user explicitly asks to override warnings, we also
override the exit code to 0.

Without this PR, it is hard to programmatically verify success, as
mentioned in #848.