Clarify clinical attribute data type error message
The error message emitted by the validator when metadata for a
clinical attribute conflicted with requirements built into the
portal, such as age being a numeric patient-level attribute,
previously only named the metadatum that was incorrect. Adding the
line and column number clarifies the meaning of the message to
people not already intimately familiar with the data file format.