Fix exception in refresh/reload of clinical tracks
Due to a change in how the state of heatmap tracks is tracked in
18c8e5e1a2e0fd70588925a92fd6c49e119ff42e, reloading an Oncoprint in
which clinical tracks are selected (by refreshing the browser or opening
the URL) made the page stop rendering with a JavaScript exception. This
commit fixes the issues by adjusting the appropriate lines to the new
style, and re-enables some commented-out logging lines.