Remove consideration about hasVariant and hasAllele
These two properties were used to help on the sorting. The non-invested
variants should have low priority in the sorting which is no longer be
the case after introducing ‘Predicted Oncogenic’. In Predicted
Oncogenic situation, the variant does not exist in OncoKB, and it may
not necessarily have alternative alleles, but it will be marked as part
of Oncogenic logic.