Suppress false-alarm PyLint errors in importer
In two lines of the Python code, PyLint makes false assumptions
about the types of variables, possibly because the code isn't phrased
in a very Pythonic way. Either way, I believe the code is correct
and the lines do not need to appear in IDEs as errors in the project.