Alter api-legacy model: pmid now String in DBStudy
The database representation for PMID is string, and the new api delivers
a string type for this field. The importer allows import of non-numeric
values into pmid, but because legacy-api is implemented to deliver this
field as Long, exceptions are thrown if any non-numeric entries are
present. To avoid these exceptions, the model (DBStudy) for the legacy
API has been altered to also return String. Dependencies on the legacy
API's representation of PMID were searched for and not found in the
cBioPortal codebase.

This addresses issue #1456